+++ sdk/monitor/azure-mgmt-monitor/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 0.9.0 (2020-04-09)
+
+**Features**
+
+  - Model AzureMonitorPrivateLinkScope has a new parameter private_endpoint_connections
+
+**Breaking changes**
+
+  - Operation PrivateLinkScopedResourcesOperations.create_or_update has a new signature
+  - Model PrivateEndpointConnection no longer has parameter tags
+  - Model PrivateLinkResource no longer has parameter tags
+  - Model ScopedResource no longer has parameter tags
+  - Model ProxyResource no longer has parameter tags
+  - Operation PrivateEndpointConnectionsOperations.create_or_update has a new signature
+  - Model ErrorResponse has a new signature
+
 ## 0.8.0 (2020-03-14)
 
 **Features**